docshell/test/test_bug968273.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=968273

If a load has redirects, reloading the page will load the page starting with the original
URI and do the redirects again.
-->
<head>
  <title>Test for Bug 968273</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>

<script type='application/javascript;version=1.7'>
  SimpleTest.waitForExplicitFinish();

  var doOneReload = true;
  var ifr;
  function checkFrame() {
    is(frames[0].performance.navigation.redirectCount, 1, "Expected rediect");
    if (doOneReload == true) {
      doOneReload = false;
      // do a reload and check that a redirect is preformed again.
      frames[0].location.reload();
    } else {
      SimpleTest.finish();
    }
  }

  ifr = document.createElement("iframe");
  ifr.src = "bug968273_redirect.html";
  ifr.onload = checkFrame;
  document.body.appendChild(ifr);

</script>
</body>
</html>
